requirements.txt
#
django>=4.0
python-dotenv
django-cors-headers
djangorestframework-simplejwt
django-ckeditor-5
# psycopg2
psycopg2-binary==2.9.3

django-extensions
getmac
django-taggit
Pillow
django-currentuser
user-agents
django-user-agents
unidecode
requests
gunicorn
django-protected-media
djangorestframework-api-key
boto3
django-storages
cryptography
gunicorn
whitenoise
dj-database-url
solana
pynacl
django-cloudinary-storage